// components/CMM4TestComponent/0.1.0.0/src/ts/layer3/CMM4TestComponentModel.interface.ts
/**
 * CMM4TestComponentModel - CMM4TestComponent Component Model Interface
 * Web4 pattern: Component model following auto-discovery patterns
 */

import { Model } from './Model.interface.js';

export interface CMM4TestComponentModel extends Model {
  uuid: string;
  name: string;
  origin: string;
  definition: string;
  createdAt: string;
  updatedAt: string;
}